ok techniques

+++ client/src/components/adminpagecomponents/showAllNamesInLIsts.js
@@ -1,5 +1,4 @@
 import React, { useContext } from 'react'
-import { useAdmin } from '../../hooks/admin.hooks'
 import { Button } from 'react-bootstrap'
 import { AdminContext } from '../../context/admin.context'
 
@@ -7,11 +6,20 @@ export const ShowAllNamesInLIsts = (props) => {
   const {deleteItem}=useContext(AdminContext)
   const {header}=props
   const componentKeyName = Object.keys(props)[0]
+  // console.log("ShowAllNamesInLIsts -> componentKeyName", componentKeyName)
   const items=props[Object.keys(props)[0]] // here array of items -> should show name field
+  
+  // console.log("ShowAllNamesInLIsts -> items", items)
 
   const deleteItemHandler =(id) => {
-    deleteItem(componentKeyName, id)
+    const answer = prompt("Really delete? write yes", 'NO')
+    if (answer.toLowerCase()==='yes') 
+      {
+        deleteItem(componentKeyName, id)
+      }
   }
+  
+  if (typeof items==='undefined') return null
     const rez = items.map((st) => (
       <div key={st._id}>
         <Button variant="outline-info" onClick={() => deleteItemHandler(st._id)}>
